Cleanup, refactor and fix tests for an initial GEM release Organize files and tests like state_machines-activerecord and get it working with * current state_machines GEM * Mongoid v4.0
