+++ lib/state_machines/integrations/mongoid/locale.rb
@@ -1,0 +1,11 @@
+{ en: {
+  mongoid: {
+    errors: {
+      messages: {
+        invalid: StateMachines::Machine.default_messages[:invalid],
+        invalid_event: StateMachines::Machine.default_messages[:invalid_event] % ['%{state}'],
+        invalid_transition: StateMachines::Machine.default_messages[:invalid_transition] % ['%{event}']
+      }
+    }
+  }
+} }
