+++ lib/state_machines/integrations/mongoid/version.rb
@@ -1,7 +1,7 @@
 module StateMachines
   module Integrations
     module Mongoid
-      VERSION = '0.0.0'
+      VERSION = '0.1.0'
     end
   end
 end
