+++ lib/state_machines/state_machines-mongoid.rb
@@ -1,0 +1,1 @@
+require 'state_machines/integrations/mongoid'
