+++ state_machines-mongoid.gemspec
@@ -23,4 +23,5 @@ Gem::Specification.new do |spec|
   spec.add_dependency 'mongoid' , '>= 4.0.0'
 
   spec.add_development_dependency 'minitest', '>= 5.4.1'
+  spec.add_development_dependency 'minitest-reporters'
 end
