+++ test/files/en.yml
@@ -1,0 +1,5 @@
+en:
+  mongoid:
+    errors:
+      messages:
+        invalid_transition: "cannot transition"
