--- test/helper.rb
@@ -1,4 +1,0 @@
-require 'minitest/reporters'
-Minitest::Reporters.use!(Minitest::Reporters::SpecReporter.new)
-require 'minitest/autorun'
-require 'state_machines/mongoid'
