+++ test/machine_with_non_column_state_attribute_undefined_test.rb
@@ -1,0 +1,33 @@
+require_relative 'test_helper'
+
+class MachineWithNonColumnStateAttributeUndefinedTest < BaseTestCase
+  def setup
+    @model = new_model do
+      def initialize
+        # Skip attribute initialization
+        @initialized_state_machines = true
+        super
+      end
+    end
+
+    @machine = StateMachines::Machine.new(@model, :status, :initial => :parked)
+    @machine.other_states(:idling)
+    @record = @model.new
+  end
+
+  def test_should_define_a_new_key_for_the_attribute
+    refute_nil @model.fields['status']
+  end
+
+  def test_should_define_a_reader_attribute_for_the_attribute
+    assert @record.respond_to?(:status)
+  end
+
+  def test_should_define_a_writer_attribute_for_the_attribute
+    assert @record.respond_to?(:status=)
+  end
+
+  def test_should_define_an_attribute_predicate
+    assert @record.respond_to?(:status?)
+  end
+end
